xray-classification/requirements.txt
flask==1.1.1
msgpack==0.6.2
numpy==1.17.4
opencv-python==4.1.2.30
pillow==6.2.1
tensorflow==2.0.0
tf-explain==0.1.0
waitress==1.3.1
